requirements.txt
blinker==1.8.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
Deprecated==1.2.14
Flask==3.0.3
fonttools==4.53.0
haversine==2.8.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
networkx==2.8.8
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pyconcorde @ git+https://github.com/jvkersch/pyconcorde@f75601d2534de2101902cefdc250f2f98f3f8a2c
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tabulate==0.8.10
tsplib95==0.7.1
tzdata==2024.1
Werkzeug==3.0.3
wrapt==1.16.0
